Add BottomTimelinePanel, HeaderPanel, and refactor InitialModal for improved date handling and UI structure
