+++ src/App.jsx
@@ -13,9 +13,9 @@ import TabsComponent from './components/TabsComponent'; // Import TabsComponent
 import NotebookLoad from './components/NotebookLoad'; // Import NotebookLoad
 import FilteredStats from './components/FilteredStats'; // Import FilteredStats
 import SideNotebook from './components/SideNotebook'; // Import SideNotebook
+import BottomTimelinePanel from './components/BottomTimelinePanel'; // Import BottomTimelinePanel
+import HeaderPanel from './components/HeaderPanel'; // Import HeaderPanel
 import * as Tabs from '@radix-ui/react-tabs';
-
-import { Calendar, LayoutDashboard, Filter, BarChart2, BookOpen } from 'lucide-react'; // Replace FontAwesome with Lucide
 
 // Lazy load non-map components
 const TimelineSlider = lazy(() => import('./components/TimelineSlider'));
@@ -25,14 +25,6 @@ const FilterForm = lazy(() => import('./components/FilterForm'));
 const LayoutForm = lazy(() => import('./components/LayoutForm'));
 const TimeGraph = lazy(() => import('./components/TimeGraph'));
 const GlobalTimeGraph = lazy(() => import('./components/GlobalTimeGraph'));
-
-const tabDefs = [
-  { key: "tab1", icon: <Calendar />, label: "Fechas" },
-  { key: "tab2", icon: <LayoutDashboard />, label: "Layout" },
-  { key: "tab3", icon: <Filter />, label: "Filtros" },
-  { key: "tab4", icon: <BarChart2 />, label: "Estadísticas" },
-  { key: "tab5", icon: <BookOpen />, label: "Cuadernos" },
-];
 
 const App = () => {
   const [fetchCedulas, setFetchCedulas] = useState(true);
@@ -171,16 +163,7 @@ const App = () => {
               <PasswordCheck onAuthenticated={() => setIsAuthenticated(true)} />
             ) : (
               <>
-                <div className="TimelineSlider">
-                  <div className="FormsContainer">
-                      <TimelineSlider />
-                      <LayoutForm />
-                  </div>
-                  <Clustering type="personas_sin_identificar" />
-                  <GlobalTimeGraph 
-                  className="GlobalTimeGraph" 
-                  onDateSelect={handleDateSelect} />
-                </div>
+                <BottomTimelinePanel />
                 <div className="ComponentToggles">
                   {Object.entries(visibleComponents).map(([key, value]) => (
                     <label key={key}>
@@ -193,157 +176,24 @@ const App = () => {
                     </label>
                   ))}
                 </div>
-                {/* Tab Header */}
-                <header
-                  style={{
-                    width: "100%",
-                    background: "rgb(245, 245, 245)",
-                    borderBottom: "1px solid rgb(221, 221, 221)",
-                    marginBottom: 16,
-                    position: "absolute",
-                    zIndex: 9999,
-                    top: 0,
-                  }}
-                >
-                  <Tabs.Root value={toolbarTab} onValueChange={setToolbarTab}>
-                    <Tabs.List
-                      style={{
-                        display: "flex",
-                        borderBottom: "1px solid #ccc",
-                        padding: 0,
-                        margin: 0,
-                        background: "transparent",
-                      }}
-                    >
-                      {tabDefs.map((tab) => (
-                        <Tabs.Trigger
-                          key={tab.key}
-                          value={tab.key}
-                          style={{
-                            flex: 1,
-                            padding: "12px 0",
-                            border: "none",
-                            borderTop:
-                              toolbarTab === tab.key
-                                ? "3px solid #007bff"
-                                : "3px solid transparent",
-                            borderBottom: "none",
-                            background: "none",
-                            outline: "none",
-                            fontWeight:
-                              toolbarTab === tab.key ? "bold" : "normal",
-                            color: toolbarTab === tab.key ? "#007bff" : "#333",
-                            cursor: "pointer",
-                            transition: "border-top 0.2s, color 0.2s",
-                            display: "flex",
-                            flexDirection: "column",
-                            alignItems: "center",
-                            gap: 2,
-                          }}
-                        >
-                          {tab.icon}
-                          <span style={{ fontSize: 12 }}>{tab.label}</span>
-                        </Tabs.Trigger>
-                      ))}
-                    </Tabs.List>
-                  </Tabs.Root>
-                </header>
-                {/* Tab Content */}
-                <div
-                  style={{
-                    top: 0,
-                    marginTop: 49,
-                    position: "absolute",
-                    zIndex: 99,
-                    width: "100%",
-                    background: "#fff",
-                  }}
-                >
-                  {/* Tab 1 */}
-                  <div
-                    style={{
-                      display: toolbarTab === "tab1" ? "block" : "none",
-                    }}
-                  >
-                    {isFormsVisible && (
-                      <div className="DateForm">
-                        <DateFormCompact
-                          handleSubmit={handleSubmit}
-                          loading={loading}
-                          fetchCedulas={fetchCedulas}
-                          setFetchCedulas={setFetchCedulas}
-                          fetchForense={fetchForense}
-                          setFetchForense={setFetchForense}
-                        />
-                        <FetchCedulas
-                          fetchCedulas={fetchCedulas}
-                          fetchId={fetchId}
-                          onFetchComplete={handleFetchComplete}
-                        />
-                        <FetchForense
-                          fetchForense={fetchForense}
-                          fetchId={fetchId}
-                          onFetchComplete={handleFetchComplete}
-                        />
-                      </div>
-                    )}
-                  </div>
-                  {/* Tab 2 */}
-                  <div
-                    style={{
-                      display: toolbarTab === "tab2" ? "block" : "none",
-                    }}
-                  >
-                    <div style={{ padding: 16 }}>
-                     <p>LayoutForm()</p>
-                    </div>
-                  </div>
-                  {/* Tab 3 */}
-                  <div
-                    style={{
-                      display: toolbarTab === "tab3" ? "block" : "none",
-                    }}
-                  >
-                    <div style={{ padding: 16 }}>
-                    {visibleComponents.filterForm && <FilterForm />}
-                    </div>
-                  </div>
-                  {/* Tab 4 */}
-                  <div
-                    style={{
-                      display: toolbarTab === "tab4" ? "block" : "none",
-                    }}
-                  >
-                    {/* Replace with your real content for Tab 4 */}
-                    <div style={{ padding: 16 }}>
-                      {visibleComponents.currentState && <FilteredStats />}
-                      {visibleComponents.violenceCases && <SemanticGraph />}
-                    </div>
-                  </div>
-                  {/* Tab 5 */}
-                  <div
-                    style={{
-                      display: toolbarTab === "tab5" ? "block" : "none",
-                    }}
-                  >
-                    {/* NotebookLoad actions in Tab 5 */}
-                    <div style={{ padding: 16 }}>
-                      <NotebookLoad
-                        saveNotesToBackend={() => {
-                          console.log("Tab5: saveNotesToBackend called");
-                        }}
-                        loadNotesFromBackend={() => {
-                          console.log("Tab5: loadNotesFromBackend called");
-                        }}
-                        listNotebooks={listNotebooksApp}
-                        isModalOpen={isNotebookModalOpen}
-                        setIsModalOpen={setIsNotebookModalOpen}
-                        notebookList={notebookList}
-                      />
-                    </div>
-                  </div>
-                </div>
-
+                <HeaderPanel 
+                  toolbarTab={toolbarTab}
+                  setToolbarTab={setToolbarTab}
+                  handleSubmit={handleSubmit}
+                  loading={loading}
+                  fetchCedulas={fetchCedulas}
+                  setFetchCedulas={setFetchCedulas}
+                  fetchForense={fetchForense}
+                  setFetchForense={setFetchForense}
+                  fetchId={fetchId}
+                  handleFetchComplete={handleFetchComplete}
+                  isFormsVisible={isFormsVisible}
+                  visibleComponents={visibleComponents}
+                  isNotebookModalOpen={isNotebookModalOpen}
+                  setIsNotebookModalOpen={setIsNotebookModalOpen}
+                  notebookList={notebookList}
+                  listNotebooksApp={listNotebooksApp}
+                />
                 <InitialModal
                   handleSubmit={handleSubmit}
                   loading={loading}
